AllInOneAPI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
google-generativeai==0.3.2
alpaca-py==0.20.3
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
cors==0.0.1
python-cors==1.7.0
